+++ src/blob/persistence/SqlBlobMetadataStore.ts
@@ -487,8 +487,8 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
     account: string,
     prefix: string = "",
     maxResults: number = DEFAULT_LIST_CONTAINERS_MAX_RESULTS,
-    marker: number | undefined
-  ): Promise<[ContainerModel[], number | undefined]> {
+    marker: string
+  ): Promise<[ContainerModel[], string | undefined]> {
     const whereQuery: any = { accountName: account };
 
     if (prefix.length > 0) {
@@ -497,8 +497,8 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
       };
     }
 
-    if (marker !== undefined) {
-      whereQuery.containerId = {
+    if (marker !== "") {
+      whereQuery.containerName = {
         [Op.gt]: marker
       };
     }
@@ -506,7 +506,7 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
     const findResult = await ContainersModel.findAll({
       limit: maxResults,
       where: whereQuery as any,
-      order: [["containerId", "ASC"]]
+      order: [["containerName", "ASC"]]
     });
 
     const leaseUpdateMapper = (model: ContainersModel) => {
@@ -517,11 +517,15 @@ export default class SqlBlobMetadataStore implements IBlobMetadataStore {
       ).sync(new ContainerLeaseSyncer(containerModel));
     };
 
-    if (findResult.length < maxResults) {
+    if (findResult.length <= maxResults) {
       return [findResult.map(leaseUpdateMapper), undefined];
     } else {
       const tail = findResult[findResult.length - 1];
-      const nextMarker = this.getModelValue<number>(tail, "containerId", true);
+      const nextMarker = this.getModelValue<string>(
+        tail,
+        "containerName",
+        true
+      );
       return [findResult.map(leaseUpdateMapper), nextMarker];
     }
   }